+++ javascript-modules/engines/eleventy-engine/lib/engine.js
@@ -6,6 +6,7 @@ import translateLiquid from './translateLiquid.js';
  */
 import unbind from './plugins/unbind.js';
 import slug from './plugins/slug-plugin.js';
+import markdownify from './plugins/markdownify.js';
 import loopContext from './plugins/loop_context.js';
 import urlFilterBuilder from './plugins/url.js';
 
@@ -21,7 +22,7 @@ export class Engine {
         this.name = options.name;
         this.files = options.files;
         this.plugins = options.plugins || [];
-        this.plugins.push(unbind, slug, loopContext);
+        this.plugins.push(unbind, slug, loopContext, markdownify);
 
         this.meta = {};
         this.info = {};